+++ Setup/InstallData.php
@@ -35,6 +35,7 @@ use Magento\Framework\Setup\InstallDataInterface;
 use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\ModuleDataSetupInterface;
 use Magento\Sales\Model\Order;
+use Magento\Framework\DB\Ddl\Table;
 
 class InstallData implements InstallDataInterface
 {
@@ -63,33 +64,25 @@ class InstallData implements InstallDataInterface
         $tableName = $installer->getConnection()->getTableName(
             'sales_order_grid'
         );
+        $columnName = 'multisafepay_status';
 
-        if ($installer->getConnection()->isTableExists($tableName)) {
+        if ($installer->getConnection()->isTableExists($installer->getTable($tableName))) {
             $salesSetup = $this->salesSetupFactory->create(
                 ['resourceName' => 'sales_setup', 'setup' => $installer]
             );
 
             $salesSetup->addAttribute(
                 Order::ENTITY,
-                'multisafepay_status',
+                $columnName,
                 [
-                    'type'     => \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
+                    'type'     => Table::TYPE_TEXT,
                     'length'   => 255,
                     'visible'  => false,
                     'nullable' => true
                 ]
             );
-
-            $installer->getConnection()->addColumn(
-                $installer->getTable($tableName),
-                'multisafepay_status',
-                [
-                    'type'    => \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                    'length'  => 255,
-                    'comment' => 'MultiSafepay status'
-                ]
-            );
         }
+
         $installer->endSetup();
     }
 }